feat(types): Move TextlintResult/TextlintMessage type to @textlint/types
