+++ packages/@textlint/kernel/src/index.ts
@@ -2,18 +2,13 @@
 export { TextlintKernel } from "./textlint-kernel";
 // Kernel Descriptor
 export * from "./descriptor/index";
-// Types
-export {
-    TextlintResult,
-    TextlintFixResult,
-    TextlintFixCommand,
-    TextlintMessage,
-    // Kernel rule/filter/plugin format
-    TextlintKernelRule,
-    TextlintKernelFilterRule,
-    TextlintKernelPlugin
-} from "./textlint-kernel-interface";
+// Kernel rule/filter/plugin format
+export { TextlintKernelRule, TextlintKernelFilterRule, TextlintKernelPlugin } from "./textlint-kernel-interface";
 
+/**
+ * Types of  textlint lint/fix result
+ */
+export { TextlintResult, TextlintFixResult, TextlintMessageFixCommand, TextlintMessage } from "@textlint/types";
 /**
  * @deprecated These types will removed in the future. Use @textlint/types instead of it
  * If you use these types in your rule, you should use @textlint/types for your rule.
