+++ packages/@textlint/kernel/src/messages/filter-duplicated-process.ts
@@ -1,6 +1,6 @@
 // LICENSE : MIT
 "use strict";
-import { TextlintMessage } from "../textlint-kernel-interface";
+import { TextlintMessage } from "@textlint/types";
 
 /**
  * @param {TextlintMessage} aMessage
