+++ packages/@textlint/kernel/src/textlint-kernel-interface.ts
@@ -91,52 +91,3 @@ export interface TextlintKernelOptions {
     // It is a value of context.getConfigBaseDir.
     configBaseDir?: string;
 }
-
-// "range" will be replaced by "text"
-export interface TextlintFixCommand {
-    text: string;
-    range: [number, number];
-}
-
-export interface TextlintMessage {
-    // See src/shared/type/MessageType.js
-    // Message Type
-    type: string;
-    // Rule Id
-    ruleId: string;
-    message: string;
-    // optional data
-    data?: any;
-    // FixCommand
-    fix?: TextlintFixCommand;
-    // location info
-    // Text -> AST TxtNode(0-based columns) -> textlint -> TextlintMessage(**1-based columns**)
-    line: number; // start with 1
-    column: number; // start with 1
-    // indexed-location
-    index: number; // start with 0
-    // Severity Level
-    // See src/shared/type/SeverityLevel.js
-    severity: number;
-}
-
-// Linting result
-export interface TextlintResult {
-    filePath: string;
-    messages: TextlintMessage[];
-}
-
-// Fixing result
-export interface TextlintFixResult {
-    filePath: string;
-    // fixed content
-    output: string;
-    // all messages = pre-applyingMessages + remainingMessages
-    // it is same with one of `TextlintResult`
-    messages: TextlintMessage[];
-    // applied fixable messages
-    applyingMessages: TextlintMessage[];
-    // original means original for applyingMessages and remainingMessages
-    // pre-applyingMessages + remainingMessages
-    remainingMessages: TextlintMessage[];
-}
