+++ packages/@textlint/types/src/Message/TextlintResult.ts
@@ -1,0 +1,48 @@
+// "range" will be replaced by "text"
+export interface TextlintMessageFixCommand {
+    text: string;
+    range: [number, number];
+}
+
+export interface TextlintMessage {
+    // See src/shared/type/MessageType.js
+    // Message Type
+    type: string;
+    // Rule Id
+    ruleId: string;
+    message: string;
+    // optional data
+    data?: any;
+    // FixCommand
+    fix?: TextlintMessageFixCommand;
+    // location info
+    // Text -> AST TxtNode(0-based columns) -> textlint -> TextlintMessage(**1-based columns**)
+    line: number; // start with 1
+    column: number; // start with 1
+    // indexed-location
+    index: number; // start with 0
+    // Severity Level
+    // See src/shared/type/SeverityLevel.js
+    severity: number;
+}
+
+// Linting result
+export interface TextlintResult {
+    filePath: string;
+    messages: TextlintMessage[];
+}
+
+// Fixing result
+export interface TextlintFixResult {
+    filePath: string;
+    // fixed content
+    output: string;
+    // all messages = pre-applyingMessages + remainingMessages
+    // it is same with one of `TextlintResult`
+    messages: TextlintMessage[];
+    // applied fixable messages
+    applyingMessages: TextlintMessage[];
+    // original means original for applyingMessages and remainingMessages
+    // pre-applyingMessages + remainingMessages
+    remainingMessages: TextlintMessage[];
+}
